Updated to v1.1
Re-tested with DiRT Rally, and added support for DiRT Rally 2.0,
and DiRT 4.

v1.0 used CreateRemoteThread to get into the target process.
Unfortunately, this only worked for 32-bit games as DirtFix.exe
was 32-bit. DiRT Rally 2.0 is a 64-bit process and required a
64-bit source process. Spawning a 64-bit helper worked, but the
heuristics of some malware scanners were starting to get twitchy.

v1.1 uses a shim DLL to get between the game and the library we
want to modify. This requires only file copies, where the game
voluntarily loads us into its process. This works for both 32-bit
and 64-bit processes, and no longer needs a watchdog process to
remain running to perform the injection -- better all round.